+++ DCManagement/functions/deployment/Install-DCChildDomain.ps1
@@ -44,7 +44,7 @@
 	.PARAMETER LogPath
 		The path where the NTDS logs should be stored.
 	
-	.PARAMETER Sysvolpath
+	.PARAMETER SysvolPath
 		The path where SYSVOL should be stored.
 	
 	.PARAMETER DatabasePath
@@ -110,7 +110,7 @@
 		$LogPath = (Get-PSFConfigValue -FullName 'DCManagement.Defaults.LogPath'),
 
 		[string]
-		$Sysvolpath = (Get-PSFConfigValue -FullName 'DCManagement.Defaults.SysvolPath'),
+		$SysvolPath = (Get-PSFConfigValue -FullName 'DCManagement.Defaults.SysvolPath'),
 
 		[string]
 		$DatabasePath = (Get-PSFConfigValue -FullName 'DCManagement.Defaults.DatabasePath'),
@@ -168,7 +168,7 @@
 				DomainMode = 'Win2012R2'
 				DatabasePath = $Configuration.DatabasePath
 				LogPath = $Configuration.LogPath
-				SysvolPath = $Configuration.Sysvol
+				SysvolPath = $Configuration.SysvolPath
 				InstallDNS = $Configuration.InstallDNS
 				SafeModeAdministratorPassword = $Configuration.SafeModeAdministratorPassword
 				NoRebootOnCompletion = $Configuration.NoRebootOnCompletion