+++ DCManagement/changelog.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
- 
+
+## ???
+
+- Fix: Install-DCChildDomain - does not respect Sysvol configuration / parameter
+- Fix: Install-DCDomainController - does not respect Sysvol configuration / parameter
+- Fix: Install-DCRootDomain - does not respect Sysvol configuration / parameter
+
 ## 1.1.8 (2020-07-03)
 
 - Fix: Install-DCDomainController will now install the DC role correctly before trying to join itself to a domain as DC